- Updated the .gitignore file - Updated /models to contain skeleton code for models. - added some server configurations to app.js. - views directory for the views will be displayed to user. - public directory is for static content by conven. - refer back to package.json to see the dependencies.
-app.use(logger('dev'));
The middleware being used is logger('dev'), which is a predefined logging middleware from the "morgan" package. It logs HTTP requests to the console in a developer-friendly format. The 'dev' option specifies a concise output format.

-app.use(express.json());
The middleware is express.json(), which parses incoming requests with JSON payloads. It makes the parsed data available in req.body.

-app.use(express.urlencoded({extended: false}));
The middleware is express.urlencoded({extended: false}), which parses incoming requests with URL-encoded payloads. The extended: false option means that the URL-encoded data is parsed with the querystring library instead of the more complex qs library.

-app.use(cookieParser());
The middleware is cookieParser(), which parses Cookie header and populates req.cookies with an object keyed by the cookie names.

-app.use(express.static(path.join(__dirname, 'public')));
The middleware is express.static, which serves static files such as images, CSS files, and JavaScript files. In this case, it serves files from the 'public' directory located in the same directory as the current script (__dirname is the directory name of the current module).